+++ src/compat/firestore/collection-group/collection-group.spec.ts
@@ -49,7 +49,7 @@ describe('AngularFirestoreCollectionGroup', () => {
   });
 
   afterEach(() => {
-    app.delete().catch(() => undefined);
+    afs.firestore.disableNetwork();
   });
 
   describe('valueChanges()', () => {