+++ src/database/database.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { FirebaseApp, provideFirebaseApp, getApp, initializeApp, deleteApp } from '@angular/fire/app';
-import { Database, provideDatabase, getDatabase, connectDatabaseEmulator } from '@angular/fire/database';
+import { Database, provideDatabase, getDatabase, connectDatabaseEmulator, goOffline } from '@angular/fire/database';
 import { COMMON_CONFIG } from '../test-config';
 import { rando } from '../utils';
 
@@ -29,7 +29,7 @@ describe('Database', () => {
     });
 
     afterEach(() => {
-        deleteApp(app).catch(() => undefined);
+        goOffline(database);
     });
 
     it('should be injectable', () => {