+++ tools/jasmine.ts
@@ -1,13 +1,11 @@
+// @ts-ignore
 import Jasmine from 'jasmine';
 
 import 'reflect-metadata';
-import 'zone.js/dist/zone-node';
-import 'zone.js/dist/zone-testing';
+import 'zone.js';
 
 import { getTestBed } from '@angular/core/testing';
 import { platformServerTesting, ServerTestingModule } from '@angular/platform-server/testing';
-
-global['globalThis'] = require('globalthis/polyfill')();
 
 // First, initialize the Angular testing environment.
 getTestBed().initTestEnvironment(
@@ -15,13 +13,13 @@ getTestBed().initTestEnvironment(
     platformServerTesting()
 );
 
-jasmine = new Jasmine();
-jasmine.loadConfig({
+const tests = new Jasmine({});
+tests.loadConfig({
     spec_dir: '.',
     spec_files: [
         'dist/out-tsc/jasmine/**/*.jasmine.js',
-        'dist/out-tsc/jasmine/**/*.spec.js'
+        'dist/out-tsc/jasmine/**/*.spec.js',
     ]
 });
 
-jasmine.execute();
+tests.execute();